requirements.dev.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
gitignoreio==0.1.5
idna==2.8
more-itertools==5.0.0
pluggy==0.8.0
py==1.7.0
pytest==4.0.2
pytest-cov==2.6.0
requests==2.21.0
six==1.12.0
urllib3==1.24.1
